Fix log in flicker issue.
On page load, AlrightEros.currentUser had no ID, so it briefly rendered the not-logged-in version. I added current_user.id into the ERB template that builds the index page. I also verified what I was pretty confident was true: you can't use this to trick the server into thinking you're a different user. Now it passes that current_user.id into the AlrightEros.currentUser as id before initializing the rest of the app.